Create the ID for records
One of the issues I had run into was re-running an import script could
duplicate entries. Elasticsearch will insert any record supplied without
an "id" field, but if an ID is supplied, it will update the record with
that ID. So now an ID is generated as the hash of the other fields in
the record. This way, if the same entry is imported a second time, it
will update the old record (noop) instead of duplicating it.